#FLOPOP-XXX: Finish Popover prototype with desktop mode simulation as Flo mac app.
