+++ FLOPopupPrototypes/Common/Popover/FLOPopoverBackgroundView.m
@@ -303,7 +303,7 @@ static CGFloat getMedianYFromRects(CGRect r1, CGRect r2) {
     CGFloat minY = NSMinY(contentRect);
     CGFloat maxY = NSMaxY(contentRect);
     
-    NSWindow *window = (self.window != nil) ? self.window : [[FLOPopoverWindowController sharedInstance] applicationWindow];
+    NSWindow *window = (self.window != nil) ? self.window : [[FLOPopoverWindow sharedInstance] appMainWindow];
     CGRect windowRect = [window convertRectFromScreen:self.popoverOrigin];
     CGRect originRect = [self convertRect:windowRect fromView:nil];
     CGFloat midOriginX = floor(getMedianXFromRects(originRect, contentRect));
@@ -403,7 +403,7 @@ static CGFloat getMedianYFromRects(CGRect r1, CGRect r2) {
 #pragma mark - Mouse events
 #pragma mark -
 - (void)mouseDown:(NSEvent *)event {
-    BOOL isFLOWindowPopover = self.window != [[FLOPopoverWindowController sharedInstance] applicationWindow];
+    BOOL isFLOWindowPopover = self.window != [[FLOPopoverWindow sharedInstance] appMainWindow];
     self.originalMouseOffset = isFLOWindowPopover ? event.locationInWindow : [self convertPoint:event.locationInWindow fromView:self.window.contentView];
     self.dragging = NO;
 }
@@ -413,7 +413,7 @@ static CGFloat getMedianYFromRects(CGRect r1, CGRect r2) {
     
     if (NSEqualSizes(self.arrowSize, NSZeroSize) && (self.shouldMovable || self.shouldDetach)) {
         if (self.dragging) {
-            BOOL isFLOWindowPopover = self.window != [[FLOPopoverWindowController sharedInstance] applicationWindow];
+            BOOL isFLOWindowPopover = self.window != [[FLOPopoverWindow sharedInstance] appMainWindow];
             
             NSPoint currentMouseOffset = isFLOWindowPopover ? event.locationInWindow : [self convertPoint:event.locationInWindow fromView:self.window.contentView];
             NSPoint difference = NSMakePoint(currentMouseOffset.x - self.originalMouseOffset.x, currentMouseOffset.y - self.originalMouseOffset.y);
@@ -431,7 +431,7 @@ static CGFloat getMedianYFromRects(CGRect r1, CGRect r2) {
 
 - (void)mouseUp:(NSEvent *)event {
     if (self.dragging) {
-        NSWindow *applicationWindow = [[FLOPopoverWindowController sharedInstance] applicationWindow];
+        NSWindow *applicationWindow = [[FLOPopoverWindow sharedInstance] appMainWindow];
         BOOL isFLOWindowPopover = self.window != applicationWindow;
         
         if ([self.delegate respondsToSelector:@selector(didPopoverMakeMovement)]) {
