+++ FLOPopupPrototypes/Common/Popover/FLOPopoverConstants.h
@@ -12,17 +12,6 @@
 #define FLO_CONST_SHOULD_USE_ANIMATION_FRAME
 #define FLO_CONST_ANIMATION_TIME_INTERVAL_STANDARD                                          0.2f
 
-#define FLO_NOTIFICATION_WINDOW_WILL_CHANGE_MODE                                            @"FLO_Notification_WindowWillChangeMode"
-#define FLO_NOTIFICATION_WINDOW_DID_CHANGE_MODE                                             @"FLO_Notification_WindowDidChangeMode"
-
-#define FLO_ENTITLEMENT_APP_IDENTIFIER_FINDER                                               @"com.apple.finder"
-#define FLO_ENTITLEMENT_APP_IDENTIFIER_SAFARI                                               @"com.apple.safari"
-
-typedef NS_ENUM(NSInteger, FLOWindowMode) {
-    FLOWindowModeNormal,
-    FLOWindowModeDesktop
-};
-
 typedef NS_ENUM(NSInteger, FLOPopoverType) {
     FLOWindowPopover,
     FLOViewPopover
