+++ FLOPopupPrototypes/Common/Popover/FLOPopoverWindowController.h
@@ -8,13 +8,13 @@
 
 #import <Cocoa/Cocoa.h>
 
-#import "FLOPopoverConstants.h"
-
 #pragma mark -
 #pragma mark - FLOPopoverWindow
 #pragma mark -
 // FLOPopoverWindow subclased of NSWindow (can see AnimatedWindow in FloProj)
 @interface FLOPopoverWindow : NSWindow
+
+@property (nonatomic, strong, readonly) NSWindow *appMainWindow;
 
 @property (nonatomic, strong, readonly) NSWindow *topWindow;
 @property (nonatomic, strong, readonly) NSView *topView;
@@ -32,19 +32,5 @@
 #pragma mark - FLOPopoverWindowController
 #pragma mark -
 @interface FLOPopoverWindowController : NSWindowController
-
-@property (nonatomic, strong, readonly) NSWindow *applicationWindow;
-@property (nonatomic, assign, readonly) FLOWindowMode windowMode;
-@property (nonatomic, assign, readonly) BOOL windowInDesktopMode;
-@property (nonatomic, assign, readonly) NSRect windowNormalFrame;
-@property (nonatomic, assign, readonly) CGFloat windowTitleBarHeight;
-
-+ (FLOPopoverWindowController *)sharedInstance;
-
-- (void)setWindowMode;
-- (void)setWindowTitleBarHeight;
-
-- (void)hideChildenWindowsOnDeactivate;
-- (void)showChildenWindowsOnActivate;
 
 @end
