+++ FLOPopupPrototypes/UI/Scenes/Base/BaseWindowController.h
@@ -8,8 +8,21 @@
 
 #import <Cocoa/Cocoa.h>
 
-#import "FLOPopoverWindowController.h"
+@interface BaseWindowController : NSWindowController
 
-@interface BaseWindowController : FLOPopoverWindowController
+@property (nonatomic, assign, readonly) FLOWindowMode windowMode;
+@property (nonatomic, assign, readonly) BOOL windowInDesktopMode;
+@property (nonatomic, assign, readonly) NSRect windowNormalFrame;
+@property (nonatomic, assign, readonly) CGFloat windowTitleBarHeight;
+
++ (BaseWindowController *)sharedInstance;
+
+- (void)setWindowMode;
+- (void)setWindowTitleBarHeight;
+
+- (void)activate;
+- (void)hideChildenWindowsOnDeactivate;
+- (void)showChildenWindowsOnActivate;
+- (void)hideOtherAppsExceptThoseInside;
 
 @end
