+++ FLOPopupPrototypes/Utils/Constants.h
@@ -9,6 +9,17 @@
 #ifndef Constants_h
 #define Constants_h
 
+#define FLO_NOTIFICATION_WINDOW_WILL_CHANGE_MODE                                            @"FLO_Notification_WindowWillChangeMode"
+#define FLO_NOTIFICATION_WINDOW_DID_CHANGE_MODE                                             @"FLO_Notification_WindowDidChangeMode"
+
+#define FLO_ENTITLEMENT_APP_IDENTIFIER_FINDER                                               @"com.apple.finder"
+#define FLO_ENTITLEMENT_APP_IDENTIFIER_SAFARI                                               @"com.apple.Safari"
+
+typedef NS_ENUM(NSInteger, FLOWindowMode) {
+    FLOWindowModeNormal,
+    FLOWindowModeDesktop
+};
+
 #define LOG_DEBUG
 
 #ifdef LOG_DEBUG
